requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12
groq
google-auth>=2.48.1
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
msal==1.31.0
requests==2.32.3
cryptography==43.0.1
chromadb==0.5.18
streamlit==1.39.0
httpx>=0.28.1
langgraph
langchain-groq
langchain-google-genai
langchain-core>=1.2.5
mcp
slack-sdk==3.27.1
PyMuPDF==1.25.3
